back/requirements.txt
aerich==0.8.2
asgiref==3.8.1
asyncpg==0.30.0
bcrypt==4.3.0
fastapi==0.115.10
PyJWT==2.10.1
tortoise-orm==0.24.2
uvicorn==0.34.0
python-dotenv==1.0.1